+++ tdmclient/atranspiler.py
@@ -13,6 +13,61 @@ import sys
 import ast
 
 
+class Context:
+
+    def __init__(self, parent_context=None, function_name=None):
+        # reference to base context for functions or None for base context itself
+        self.parent_context = parent_context
+
+        # function name, to be used as the namespace of local variables
+        self.function_name = function_name
+
+        # size of local variables
+        self.var = {}
+
+        # set of variables declared as global
+        self.global_var = set()
+
+        self.has_return_val = False
+        self.tmp_req = 0
+
+    def is_global(self, name):
+        return name not in self.var or name in ATranspiler.PREDEFINED_VARIABLES
+
+    def var_str(self, name):
+        return name if self.function_name is None or self.is_global(name) else f"_{self.function_name}_{name}"
+
+    def tmp_var_str(self, index):
+        # _tmp is always local to avoid interference with caller's
+        name = "_tmp" if self.function_name is None else f"_{self.function_name}__tmp"
+        return f"{name}[{index}]"
+
+    def declare_global(self, name):
+        if name in self.var:
+            raise Exception(f"Variable {name} declared global after being assigned to")
+        self.global_var.add(name)
+
+    def declare_var(self, name, size=None):
+        var = self.parent_context.var if name in self.global_var and self.parent_context is not None else self.var
+        if name in var:
+            if (size != var[name] or
+                name in ATranspiler.PREDEFINED_VARIABLES and size != ATranspiler.PREDEFINED_VARIABLES[name]):
+                raise Exception(f"Incompatible sizes for list assignment to {name}")
+        else:
+            var[name] = size
+
+    def var_declarations(self):
+        return "".join([
+            f"var {self.var_str(name)}{f'[{self.var[name]}]' if self.var[name] is not None else ''}\n"
+            for name in self.var
+            if name not in ATranspiler.PREDEFINED_VARIABLES
+        ])
+
+    def request_tmp(self, n):
+        if n > self.tmp_req:
+            self.tmp_req = n
+            self.var["_tmp"] = self.tmp_req
+
 class ATranspiler:
 
     PREDEFINED_VARIABLES = {
@@ -117,14 +172,13 @@ class ATranspiler:
     PRI_UNARY_MINUS = 17
     PRI_HIGH = 100
 
-    def compile_expr(self, node, priority_container=PRI_LOW, tmp_offset=0):
+    def compile_expr(self, node, context, priority_container=PRI_LOW, tmp_offset=0):
         """Compile an expression or subexpression.
         Return the expression, additional statements to calculate auxiliary values,
-        total requirements of temporary array tmp[], and whether result is boolean.
+        and whether result is boolean.
         """
         code = None
         aux_statements = ""
-        tmp_req = tmp_offset
         priority = self.PRI_HIGH
         is_boolean = False
 
@@ -144,9 +198,9 @@ class ATranspiler:
                 ast.RShift: (">>", self.PRI_SHIFT),
                 ast.Sub: ("-", self.PRI_ADD),
             }[type(op)]
-            left, aux_st, tmp_req, _ = self.compile_expr(node.left, priority, tmp_req)
+            left, aux_st, _ = self.compile_expr(node.left, context, priority, tmp_offset)
             aux_statements += aux_st
-            right, aux_st, tmp_req, _ = self.compile_expr(node.right, priority, tmp_req)
+            right, aux_st, _ = self.compile_expr(node.right, context, priority, tmp_offset)
             aux_statements += aux_st
             code = f"{left} {op_str} {right}"
         elif isinstance(node, ast.BoolOp):
@@ -154,11 +208,9 @@ class ATranspiler:
             # result is not pretty b/c of aseba's idea of what's acceptable
             op = node.op
             cmp = "!=" if isinstance(op, ast.And) else "=="
-            tmp_offset = tmp_req
-            tmp_req += 1
+            context.request_tmp(tmp_offset + 1)
             for i in range(len(node.values)):
-                value, aux_st, tmp_req1, is_value_boolean = self.compile_expr(node.values[i], self.PRI_ASSIGN, tmp_offset + 1)
-                tmp_req = max(tmp_req, tmp_req1)
+                value, aux_st, is_value_boolean = self.compile_expr(node.values[i], context, self.PRI_ASSIGN, tmp_offset + 1)
                 # store value into tmp[tmp_offset]
                 aux_statements += aux_st
                 if is_value_boolean:
@@ -169,16 +221,16 @@ tmp[{tmp_offset}] = 0
 end
 """
                 else:
-                    aux_statements += f"""tmp[{tmp_offset}] = {value}
+                    aux_statements += f"""{context.tmp_var_str(tmp_offset)} = {value}
 """
                 # continue evaluating terms if true (and) or false (or)
                 if i + 1 < len(node.values):
-                    aux_statements += f"""if tmp[{tmp_offset}] {cmp} 0 then
+                    aux_statements += f"""if {context.tmp_var_str(tmp_offset)} {cmp} 0 then
 """
             for i in range(len(node.values) - 1):
                 aux_statements += """end
 """
-            code = f"tmp[{tmp_offset}]"
+            code = context.tmp_var_str(tmp_offset)
             is_boolean = False
         elif isinstance(node, ast.Call):
             # a very few set of functions
@@ -188,10 +240,10 @@ end
             if fun_name == "abs":
                 if len(node.args) != 1:
                     raise Exception("Wrong number of arguments for abs")
-                code, aux_st, tmp_req, is_boolean = self.compile_expr(node.args[0], self.PRI_COMMA, tmp_req)
+                code, aux_st, is_boolean = self.compile_expr(node.args[0], context, self.PRI_COMMA, tmp_offset)
                 aux_statements += aux_st
                 code = f"abs({code})"
-                return code, aux_statements, tmp_req, False
+                return code, aux_statements, False
             else:
                 raise Exception(f"Unknown function {fun_name}")
         elif isinstance(node, ast.Compare):
@@ -207,9 +259,9 @@ end
                 ast.NotEq: "!=",
             }[type(op)]
             priority = self.PRI_COMPARISON
-            left, aux_st, tmp_req, _ = self.compile_expr(node.left, self.PRI_NUMERIC, tmp_req)
+            left, aux_st, _ = self.compile_expr(node.left, context, self.PRI_NUMERIC, tmp_offset)
             aux_statements += aux_st
-            right, aux_st, tmp_req, _ = self.compile_expr(node.comparators[0], self.PRI_NUMERIC, tmp_req)
+            right, aux_st, _ = self.compile_expr(node.comparators[0], context, self.PRI_NUMERIC, tmp_offset)
             aux_statements += aux_st
             if op_str is None:
                 raise Exception(f"Comparison op {ast.dump(op)} not implemented")
@@ -226,45 +278,45 @@ end
             if priority_container > self.PRI_ASSIGN:
                 raise Exception("List not supported in expression")
             for el in node.elts:
-                el_code, aux_st, tmp_req, is_boolean = self.compile_expr(el, self.PRI_NUMERIC, tmp_req)
+                el_code, aux_st, is_boolean = self.compile_expr(el, context, self.PRI_NUMERIC, tmp_offset)
                 aux_statements += aux_st
                 if code is None:
                     code = "[" + el_code
                 else:
                     code += ", " + el_code
             code += "]"
-            return code, aux_statements, tmp_req, False
+            return code, aux_statements, False
         elif isinstance(node, ast.Name):
-            code = self.decode_attr(node)
+            code = context.var_str(node.id)
         elif isinstance(node, ast.Subscript):
             name = self.decode_attr(node.value)
             index = node.slice.value
-            index_value, aux_st, tmp_req, is_index_boolean = self.compile_expr(index, self.PRI_NUMERIC, tmp_req)
+            index_value, aux_st, is_index_boolean = self.compile_expr(index, context, self.PRI_NUMERIC, tmp_offset)
             if is_index_boolean:
                 aux_st += f"""if {index_value} then
-tmp[{tmp_req}] = 1
+{context.tmp_var_str(tmp_offset)} = 1
 else
-tmp[{tmp_req}] = 0
+{context.tmp_var_str(tmp_offset)} = 0
 end
 """
-                index_value = f"tmp[{tmp_req}]"
-                tmp_req += 1
+                index_value = context.tmp_var_str(tmp_req)
+                context.request_tmp(tmp_offset + 1)
             aux_statements += aux_st
             code = f"{name}[{index_value}]"
         elif isinstance(node, ast.UnaryOp):
             op = node.op
             if isinstance(op, ast.UAdd):
                 priority = self.PRI_ADD
-                code, aux_st, tmp_req, is_boolean = self.compile_expr(node.operand, priority, tmp_req)
+                code, aux_st, is_boolean = self.compile_expr(node.operand, context, priority, tmp_offset)
                 aux_statements += aux_st
             elif isinstance(op, ast.Invert):
                 priority = self.PRI_BINARY_NOT
-                operand, aux_st, tmp_req, is_boolean = self.compile_expr(node.operand, priority, tmp_req)
+                operand, aux_st, is_boolean = self.compile_expr(node.operand, context, priority, tmp_offset)
                 aux_statements += aux_st
                 code = "~" + operand
             elif isinstance(op, ast.Not):
                 priority = self.PRI_LOGICAL_NOT
-                operand, aux_st, tmp_req, is_boolean = self.compile_expr(node.operand, priority, tmp_req)
+                operand, aux_st, is_boolean = self.compile_expr(node.operand, context, priority, tmp_offset)
                 if is_boolean:
                     code = "not " + operand
                 else:
@@ -272,7 +324,7 @@ end
                     is_boolean = True
             elif isinstance(op, ast.USub):
                 priority = self.PRI_UNARY_MINUS
-                operand, aux_st, tmp_req, is_boolean = self.compile_expr(node.operand, priority, tmp_req)
+                operand, aux_st, is_boolean = self.compile_expr(node.operand, context, priority, tmp_offset)
                 aux_statements += aux_st
                 code = "-" + operand
             else:
@@ -284,17 +336,17 @@ end
             # work around aseba's idea of what's acceptable
             # (no boolean in arithmetic subexpressions or variables)
             aux_statements += f"""if {code} then
-tmp[{tmp_offset}] = 1
+{context.tmp_var_str(tmp_offset)} = 1
 else
-tmp[{tmp_offset}] = 0
+{context.tmp_var_str(tmp_offset)} = 0
 end
 """
-            tmp_req = max(tmp_req, tmp_offset + 1)
-            return f"tmp[{tmp_offset}]", aux_statements, tmp_req, False
+            context.request_tmp(tmp_offset + 1)
+            return context.tmp_var_str(tmp_offset), aux_statements, False
         elif priority < priority_container:
-            return "(" + code + ")", aux_statements, tmp_req, is_boolean
+            return "(" + code + ")", aux_statements, is_boolean
         else:
-            return code, aux_statements, tmp_req, is_boolean
+            return code, aux_statements, is_boolean
 
     def decode_target(self, target):
         """Decode an assignment target and return variable name (possibly dotted) and
@@ -307,53 +359,55 @@ end
         name = self.decode_attr(target)
         return name, index
 
-    def compile_node(self, node, tmp_req=0, var0=None):
+    def compile_node(self, node, context, tmp_offset=0, var0=None):
         code = ""
         if isinstance(node, ast.Assign):
             if len(node.targets) != 1:
                 raise Exception("Unsupported assignment to multiple targets")
             target, index = self.decode_target(node.targets[0])
+            target_str = context.var_str(target)
             if isinstance(node.value, ast.List):
                 if index is not None:
                     raise Exception("List assigned to indexed variable")
-                value, aux_statements, tmp_req, is_boolean = self.compile_expr(node.value, self.PRI_ASSIGN, tmp_req)
+                value, aux_statements, is_boolean = self.compile_expr(node.value, context, self.PRI_ASSIGN, tmp_offset)
             else:
-                value, aux_statements, tmp_req, is_boolean = self.compile_expr(node.value, self.PRI_NUMERIC, tmp_req)
+                value, aux_statements, is_boolean = self.compile_expr(node.value, context, self.PRI_NUMERIC, tmp_offset)
             code += aux_statements
             if index is not None:
-                index_value, aux_statements, tmp_req, is_index_boolean = self.compile_expr(index, self.PRI_NUMERIC, tmp_req)
+                index_value, aux_statements, is_index_boolean = self.compile_expr(index, context, self.PRI_NUMERIC, tmp_offset)
                 code += aux_statements
                 if is_index_boolean:
                     code += f"""if {index_value} then
-tmp[{tmp_req}] = 1
+{context.tmp_var_str(tmp_offset)} = 1
 else
-tmp[{tmp_req}] = 0
+{context.tmp_var_str(tmp_offset)} = 0
 end
 """
-                    index_value = "tmp[{tmp_req}]"
-                    tmp_req += 1
-                target += "[" + index_value + "]"
+                    index_value = context.tmp_var_str(tmp_offset)
+                    context.request_tmp(tmp_offset + 1)
+                target_str += "[" + index_value + "]"
             if is_boolean:
                 # convert boolean to number
                 code += f"""if {value} then
-{target} = 1
+{target_str} = 1
 else
-{target} = 0
+{target_str} = 0
 end
 """
-                tmp_req = max(tmp_req, 1)
             else:
-                code += f"{target} = {value}\n"
+                code += f"{target_str} = {value}\n"
             if isinstance(node.value, ast.List):
                 # var = [...]
                 target_size = len(node.value.elts)
+                context.declare_var(target, target_size)
             elif isinstance(node.value, ast.Name):
                 # var1 = var2: inherit size
                 name_right = self.decode_attr(node.value)
                 target_size = var0[name_right] if var0 is not None and name_right in var0 else None
-            else:
-                target_size = None
-            return code, {target: target_size} if index is None else {}, tmp_req
+                context.declare_var(target, target_size)
+            elif index is None:
+                context.declare_var(target)
+            return code
         elif isinstance(node, ast.AugAssign):
             op_str = {
                 ast.Add: "+",
@@ -368,39 +422,39 @@ end
                 ast.Sub: "-",
             }[type(node.op)]
             target, index = self.decode_target(node.target)
-            value, aux_statements, tmp_req, is_boolean = self.compile_expr(node.value, self.PRI_NUMERIC, tmp_req)
+            target_str = context.var_str(context)
+            value, aux_statements, is_boolean = self.compile_expr(node.value, context, self.PRI_NUMERIC, tmp_offset)
             code += aux_statements
             if index is not None:
-                index_value, aux_statements, tmp_req, is_index_boolean = self.compile_expr(index, self.PRI_NUMERIC, tmp_req)
+                index_value, aux_statements, is_index_boolean = self.compile_expr(index, context, self.PRI_NUMERIC, tmp_offset)
                 code += aux_statements
                 if is_index_boolean:
                     code += f"""if {index_value} then
-tmp[{tmp_req}] = 1
+{context.tmp_var_str(tmp_offset)} = 1
 else
-tmp[{tmp_req}] = 0
+{context.tmp_var_str(tmp_offset)} = 0
 end
 """
-                    index_value = "tmp[{tmp_req}]"
-                    tmp_req += 1
+                    index_value = context.tmp_var_str(tmp_offset)
+                    context.request_tmp(tmp_offset + 1)
                 target += "[" + index_value + "]"
             if is_boolean:
                 # convert boolean to number
                 code += f"""if {value} then
-{target} {op_str}= 1
+{target_str} {op_str}= 1
 else
-{target} {op_str}= 0
+{target_str} {op_str}= 0
 end
 """
-                tmp_req = max(tmp_req, 1)
             else:
-                code += f"{target} {op_str}= {value}\n"
-            return code, {}, tmp_req
+                code += f"{target_str} {op_str}= {value}\n"
+            return code
         elif isinstance(node, ast.Expr):
             # plain expression without assignment
             expr = node.value
             # hard-coded ... (ellipsis, alias of None, synonym of pass)
             if isinstance(expr, ast.Ellipsis):
-                return "", {}, tmp_req
+                return ""
             # hard-coded emit(name, params...)
             if isinstance(expr, ast.Call) and isinstance(expr.func, ast.Name):
                 if expr.func.id == "emit":
@@ -413,20 +467,18 @@ end
                     code = f"emit {event_name}"
                     aux_statements = ""
                     if len(expr.args) > 1:
-                        tmp_req0 = tmp_req
                         for i in range(len(expr.args) - 1):
-                            value, aux_st, tmp_req1, is_boolean = self.compile_expr(expr.args[1 + i], self.PRI_NUMERIC, tmp_req0)
+                            value, aux_st, is_boolean = self.compile_expr(expr.args[1 + i], context, self.PRI_NUMERIC, tmp_offset)
                             aux_statements += aux_st
                             code += " [" if i == 0 else ", "
                             code += value
-                            tmp_req = max(tmp_req, tmp_req1)
                         code += "]"
                     code = aux_statements + code
-                    return code, {}, tmp_req
+                    return code
             # parse expression
-            value, aux_statements, tmp_req, is_boolean = self.compile_expr(expr, self.PRI_NUMERIC, tmp_req)
+            value, aux_statements, is_boolean = self.compile_expr(expr, context, self.PRI_NUMERIC, tmp_offset)
             # ignore it because nothing can cause side effects now
-            return "", {}, tmp_req
+            return ""
         elif isinstance(node, ast.For):
             # for var in range(...): ...
             if not isinstance(node.target, ast.Name):
@@ -438,140 +490,117 @@ end
                 raise Exception("for loop with unsupported iterator (not range)")
             range_args = node.iter.args
             target = self.decode_attr(node.target)
-            var = {target: None}
+            target_str = context.var_str(target)
+            context.declare_var(target)
             if len(range_args) == 1:
                 # for var in range(a): ...
                 # stores limit a in tmp[tmp_offset]
-                tmp_offset = tmp_req
-                value, aux_statements, tmp_req, is_boolean = self.compile_expr(range_args[0], self.PRI_NUMERIC, tmp_offset)
-                tmp_req = max(tmp_req, tmp_offset + 1)
+                value, aux_statements, is_boolean = self.compile_expr(range_args[0], context, self.PRI_NUMERIC, tmp_offset)
+                context.request_tmp(tmp_offset + 1)
                 code += aux_statements
-                code += f"""{target} = 0
-tmp[{tmp_offset}] = {value}
-while {target} < tmp[{tmp_offset}] do
+                code += f"""{target_str} = 0
+{context.tmp_var_str(tmp_offset)} = {value}
+while {target_str} < {context.tmp_var_str(tmp_offset)} do
 """
             elif len(range_args) == 2:
                 # for var in range(a, b)
                 # stores limit b in tmp[tmp_offset]
-                tmp_offset = tmp_req
-                value, aux_statements, tmp_req, is_boolean = self.compile_expr(range_args[0], self.PRI_NUMERIC, tmp_offset)
-                tmp_req = max(tmp_req, tmp_offset + 1)
+                value, aux_statements, is_boolean = self.compile_expr(range_args[0], context, self.PRI_NUMERIC, tmp_offset)
                 code += aux_statements
                 code += f"""{target} = {value}
 """
-                value, aux_statements, tmp_req1, is_boolean = self.compile_expr(range_args[1], self.PRI_NUMERIC, tmp_offset)
-                tmp_req = max(tmp_req, tmp_req1)
+                value, aux_statements, is_boolean = self.compile_expr(range_args[1], context, self.PRI_NUMERIC, tmp_offset)
+                context.request_tmp(tmp_offset + 1)
                 code += aux_statements
-                code += f"""tmp[{tmp_offset}] = {value}
-while {target} < tmp[{tmp_offset}] do
+                code += f"""{context.tmp_var_str(tmp_offset)} = {value}
+while {target_str} < {context.tmp_var_str(tmp_offset)} do
 """
             else:
                 # for var in range(a, b, c)
                 # stores limit b in tmp[tmp_offset] and step c in tmp[tmp_offset+1]
-                tmp_offset = tmp_req
-                value, aux_statements, tmp_req, is_boolean = self.compile_expr(range_args[0], self.PRI_NUMERIC, tmp_offset)
-                tmp_req = max(tmp_req, tmp_offset + 2)
+                value, aux_statements, is_boolean = self.compile_expr(range_args[0], context, self.PRI_NUMERIC, tmp_offset)
                 code += aux_statements
                 code += f"""{target} = {value}
 """
-                value, aux_statements, tmp_req1, is_boolean = self.compile_expr(range_args[1], self.PRI_NUMERIC, tmp_offset)
-                tmp_req = max(tmp_req, tmp_req1)
+                value, aux_statements, is_boolean = self.compile_expr(range_args[1], context, self.PRI_NUMERIC, tmp_offset)
+                context.request_tmp(tmp_offset + 2)
                 code += aux_statements
-                code += f"""tmp[{tmp_offset}] = {value}
+                code += f"""{context.tmp_var_str(tmp_offset)} = {value}
 """
-                value, aux_statements, tmp_req1, is_boolean = self.compile_expr(range_args[2], self.PRI_NUMERIC, tmp_offset + 1)
-                tmp_req = max(tmp_req, tmp_req1)
+                value, aux_statements, is_boolean = self.compile_expr(range_args[2], context, self.PRI_NUMERIC, tmp_offset + 1)
                 code += aux_statements
-                code += f"""tmp[{tmp_offset + 1}] = {value}
-while {target} < tmp[{tmp_offset}] do
+                code += f"""{context.tmp_var_str(tmp_offset + 1)} = {value}
+while {target_str} < {context.tmp_var_str(tmp_offset)} do
 """
-            body, var1, tmp_req1 = self.compile_node_array(node.body, tmp_offset, var0=var0)
+            body = self.compile_node_array(node.body, context, tmp_offset + 1)
             code += body
-            self.check_var_size(var, var1)
-            var = {**var, **var1}
-            tmp_req = max(tmp_req, tmp_req1)
             if len(range_args) <= 2:
                 # just increment target
-                code += f"""{target}++
+                code += f"""{target_str}++
 """
             else:
                 # increment target by step
-                code += f"""{target} += tmp[{tmp_offset + 1}]
+                code += f"""{target_str} += {context.tmp_var_str(tmp_offset + 1)}
 """
             code += """end
 """
             if node.orelse is not None and len(node.orelse) > 0:
                 # else clause always executed b/c break is not supported
-                body, var1, tmp_req1 = self.compile_node_array(node.orelse, tmp_offset, var0=var0)
+                body = self.compile_node_array(node.orelse, context, tmp_offset)
                 code += body
-                self.check_var_size(var, var1)
-                var = {**var, **var1}
-                tmp_req = max(tmp_req, tmp_req1)
-            return code, var, tmp_req
+            return code
         elif isinstance(node, ast.Global):
-            # ignored, but should be used in functions for forward compatibility
-            return "", {}, tmp_req
+            for name in node.names:
+                context.declare_global(name)
+            return ""
         elif isinstance(node, ast.If):
-            tmp_offset = tmp_req
-            test_value, aux_statements, tmp_req, is_boolean = self.compile_expr(node.test, self.PRI_LOW, tmp_req)
+            test_value, aux_statements, is_boolean = self.compile_expr(node.test, context, self.PRI_LOW, tmp_offset)
             code += aux_statements
             code += f"""if {test_value}{"" if is_boolean else " != 0"} then
 """
-            body, var, tmp_req1 = self.compile_node_array(node.body, tmp_offset, var0=var0)
+            body = self.compile_node_array(node.body, context, tmp_offset)
             code += body
-            tmp_req = max(tmp_req, tmp_req1)
             while len(node.orelse) == 1 and isinstance(node.orelse[0], ast.If):
                 # "if" node as single element of orelse: elif
                 node = node.orelse[0]
-                test_value, aux_statements, tmp_req, is_boolean = self.compile_expr(node.test, self.PRI_LOW, tmp_req)
+                test_value, aux_statements, is_boolean = self.compile_expr(node.test, context, self.PRI_LOW, tmp_offset)
                 code += aux_statements
                 code += f"""elseif {test_value}{"" if is_boolean else " != 0"} then
 """
-                body, var1, tmp_req1 = self.compile_node_array(node.body, tmp_offset, var0=var0)
+                body = self.compile_node_array(node.body, context, tmp_offset)
                 code += body
-                self.check_var_size(var, var1)
-                var = {**var, **var1}
-                tmp_req = max(tmp_req, tmp_req1)
             if len(node.orelse) > 0:
                 # anything else in orelse: else
                 code += """else
 """
-                body, var1, tmp_req1 = self.compile_node_array(node.orelse, tmp_offset, var0=var0)
+                body = self.compile_node_array(node.orelse, context, tmp_offset)
                 code += body
-                self.check_var_size(var, var1)
-                var = {**var, **var1}
-                tmp_req = max(tmp_req, tmp_req1)
             code += """end
 """
-            return code, var, tmp_req
+            return code
         elif isinstance(node, ast.Pass):
-            return "", {}, tmp_req
+            return ""
         elif isinstance(node, ast.Return):
             if node.value is not None:
                 raise Exception("Unsupported value in return statement")
             code += """return
 """
-            return code, {}, tmp_req
+            return code
         elif isinstance(node, ast.While):
-            tmp_offset = tmp_req
-            test_value, aux_statements, tmp_req, is_boolean = self.compile_expr(node.test, self.PRI_LOW, tmp_req)
+            test_value, aux_statements, is_boolean = self.compile_expr(node.test, context, self.PRI_LOW, tmp_offset)
             code += aux_statements
             code += f"""while {test_value}{"" if is_boolean else " != 0"} do
 """
-            body, var, tmp_req1 = self.compile_node_array(node.body, tmp_offset, var0=var0)
+            body = self.compile_node_array(node.body, context, tmp_offset)
             code += body
-            tmp_req = max(tmp_req, tmp_req1)
             code += aux_statements  # to evaluate condition
             code += """end
 """
             if node.orelse is not None and len(node.orelse) > 0:
                 # else clause always executed b/c break is not supported
-                body, var1, tmp_req1 = self.compile_node_array(node.orelse, tmp_offset, var0=var0)
+                body = self.compile_node_array(node.orelse, context, tmp_offset)
                 code += body
-                self.check_var_size(var, var1)
-                var = {**var, **var1}
-                tmp_req = max(tmp_req, tmp_req1)
-            return code, var, tmp_req
+            return code
         else:
             raise Exception(f"Node {ast.dump(node)} not implemented")
 
@@ -582,17 +611,12 @@ while {target} < tmp[{tmp_offset}] do
                 name in ATranspiler.PREDEFINED_VARIABLES and var_new[name] != ATranspiler.PREDEFINED_VARIABLES[name]):
                 raise Exception(f"Incompatible sizes for list assignment to {name}")
 
-    def compile_node_array(self, node_array, tmp_req=0, var0=None):
+    def compile_node_array(self, node_array, context, tmp_offset=0):
         code = ""
-        var = var0 or {}
-        tmp_req0 = tmp_req
         for node in node_array:
-            c, v, tmp_req1 = self.compile_node(node, tmp_req0, var0=var)
+            c = self.compile_node(node, context, tmp_offset)
             code += c
-            self.check_var_size(var, v)
-            var = {**var, **v}
-            tmp_req = max(tmp_req, tmp_req1)
-        return code, var, tmp_req
+        return code
 
     def transpile(self):
         self.ast = ast.parse(self.src)
@@ -601,28 +625,28 @@ while {target} < tmp[{tmp_offset}] do
         self.output_src = ""
 
         # top-level code
-        self.output_src, var, tmp_req = self.compile_node_array(top_code)
+        context_top = Context()
+        self.output_src = self.compile_node_array(top_code, context_top)
 
         # onevent handlers
-        tmp_offset = tmp_req
+        context_fun = {}
         for event_name in event_handlers:
-            event_output_src, var1, tmp_req1 = self.compile_node_array(event_handlers[event_name].body, tmp_offset, var0=var)
-            tmp_req = max(tmp_req, tmp_req1)
-            self.check_var_size(var, var1)
-            var = {**var, **var1}
+            context_fun[event_name] = Context(context_top, event_name)
+            # first pass to gather local variables into context_fun[.] (not declared global, but assigned to)
+            event_output_src = self.compile_node_array(event_handlers[event_name].body, context_fun[event_name])
+            # second pass to produce transpiled code with correct local variable names
+            event_output_src = self.compile_node_array(event_handlers[event_name].body, context_fun[event_name])
             self.output_src += f"""
 onevent {event_name}
 """ + event_output_src
 
         # variable declarations
-        if tmp_req > 0:
-            var["tmp"] = tmp_req
-        if len(var) > 0:
-            var_decl = "".join([
-                f"var {v}{f'[{var[v]}]' if var[v] is not None else ''}\n"
-                for v in var
-                if v not in self.PREDEFINED_VARIABLES
-            ])
+        var_decl = context_top.var_declarations()
+        var_decl += "".join([
+            context_fun[event_name].var_declarations()
+            for event_name in event_handlers
+        ])
+        if len(var_decl) > 0:
             self.output_src = var_decl + "\n" + self.output_src
 
     @staticmethod